[FAB-5031] Fix e2e template typo
There is a rogue network section which is indented at the top level of
the docker-compose-e2e-template.yaml file.  This has been reported by
users to cause errors when they attempt to tweak the byfn setup.

As best as I can tell, this network section is intended to be defined at
the top level and appears to have no adverse affect on the byfn.

Change-Id: I40b937d1f352af079754a84860bce025ec61594e
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>