[FAB-8503] Prevent CLI container from exiting
The pattern of the byfn scripts and of our tutorial is to start the CLI
container, then docker exec commands against it.  For some reason, we
only start the CLI container with a command of 'sleep 1000' which causes
it exit after this time.

This causes headaches of having to restart the CLI container and is only
saving the overhead of one idle bash process.  From a usability
it seems far better to leave an idle bash process so that we can be
assured the container is still running.

Change-Id: I8d5c860b89226a28f15d42f5a19e42b923bfa8c1
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>