[FAB-8568] BYFN: Fix IMAGE_TAG for couchdb
As of 1.1 alpha timeframe, 3rd party images do not follow the
same docker IMAGE_TAG versioning as the other Fabric docker images.

Need to remove IMAGE_TAG from couchdb images, and default to
latest couchdb image, which is how it worked prior to
the -i introduction.

This change needs to be made in master branch but not release branch,
since the IMAGE_TAGs are aligned in 1.0.x release.

Change-Id: Iaca24ea7ed1fc6319738dcf59f665da9259d2b9a
Signed-off-by: David Enyeart <enyeart@us.ibm.com>