FAB-8612 Cleanup eyfn/byfn with --remove-orphans
There are currently multiple docker-compose downs going on in the
first-network samples.  These are in an attempt to catch different
configurations of the same compose network.  The flag --remove-orphans
was created to allow tearing down components in a composition even when
there is a mismatch in the compose definition, so this cleans the down
scripts up a bit.

Change-Id: Ied6a06e2ebc21053993e2a6b39b6fbd7fffdfe02
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>